WORKINGBITCHES

+++ setting_nodes_utils_2.c
@@ -6,7 +6,7 @@
 /*   By: salowie <salowie@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 12:52:34 by Sarah             #+#    #+#             */
-/*   Updated: 2023/08/30 13:01:37 by salowie          ###   ########.fr       */
+/*   Updated: 2023/08/30 15:59:31 by salowie          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ void	find_target_node(t_list **head_a, t_list **head_b)
 	t_list	*target_node;
 	t_list	*a;
 	t_list	*b;
-	int		biggest_value;
+	long		biggest_value;
 
 	b = *head_b;
 	while (b)
 	{
 		a = *head_a;
-		biggest_value = INT_MAX;
+		biggest_value = LONG_MAX;
 		while (a)
 		{
 			if ((a->c > b->c) && a->c < biggest_value)
@@ -34,7 +34,7 @@ void	find_target_node(t_list **head_a, t_list **head_b)
 			}
 			a = a->next;
 		}
-		if (biggest_value == INT_MAX)
+		if (biggest_value == LONG_MAX)
 			b->target = find_smallest(head_a, b->c);
 		else
 			b->target = target_node;
